상가 삭제 시 정말로 지우시겠습니까 메시지

+++ android/app/src/main/java/org/tensorflow/lite/examples/detection/coronaCounter/view/EditStore.kt
@@ -1,5 +1,6 @@
 package org.tensorflow.lite.examples.detection.coronaCounter.view
 
+import android.app.AlertDialog
 import android.content.Context
 import android.os.Bundle
 import android.util.Log
@@ -89,28 +90,39 @@ class EditStore : Fragment() {
         binding.shopName.text = shop.shopName
         deleteButton = binding.deleteButton
         deleteButton.setOnClickListener {
-            if (shop.equals(sharedViewModel.mainShop.value)){
-                Toast.makeText(getActivity(), "현재 이용중인 상가는 삭제하실 수 없습니다", Toast.LENGTH_SHORT).show();
-                Log.d(TAG,"현재 이용중인 상가는 삭제하실 수 없습니다")
-            }
-            else{
 
-                // val shop = sharedViewModel.shop.value!!
-            Log.d(TAG,"shop delete" + shop.toString())
-            lifecycleScope.launch{
-                val succeed = sharedViewModel.deleteShop(shop)
-                if (succeed){
-                    Toast.makeText(getActivity(), "상가 삭제 성공", Toast.LENGTH_SHORT).show();
-                    sharedViewModel.fetchShops()
-                    val action = EditStoreDirections.actionEditStoreToMyPage()
-                    view.findNavController().navigate(action)
-                    Log.d(TAG,"${shop.toString()} deleted")
-                }else{
-                    Toast.makeText(getActivity(), "상가 삭제 실패", Toast.LENGTH_SHORT).show();
-                    Log.d(TAG,"delete failed ${shop.toString()}")
+            val builder = AlertDialog.Builder(mycontext)
+            builder.setMessage("정말로 지우시겠습니까?").setCancelable(false).setPositiveButton("Yes"){dialog, id ->
+                if (shop.equals(sharedViewModel.mainShop.value)){
+                    Toast.makeText(getActivity(), "현재 이용중인 상가는 삭제하실 수 없습니다", Toast.LENGTH_SHORT).show();
+                    Log.d(TAG,"현재 이용중인 상가는 삭제하실 수 없습니다")
                 }
+                else{
+
+                    // val shop = sharedViewModel.shop.value!!
+                    Log.d(TAG,"shop delete" + shop.toString())
+                    lifecycleScope.launch{
+                        val succeed = sharedViewModel.deleteShop(shop)
+                        if (succeed){
+                            Toast.makeText(getActivity(), "상가 삭제 성공", Toast.LENGTH_SHORT).show();
+                            sharedViewModel.fetchShops()
+                            val action = EditStoreDirections.actionEditStoreToMyPage()
+                            view.findNavController().navigate(action)
+                            Log.d(TAG,"${shop.toString()} deleted")
+                        }else{
+                            Toast.makeText(getActivity(), "상가 삭제 실패", Toast.LENGTH_SHORT).show();
+                            Log.d(TAG,"delete failed ${shop.toString()}")
+                        }
+                    }
+                }
+            }.setNegativeButton("No") { dialog, id ->
+                // Dismiss the dialog
+                dialog.dismiss()
             }
-            }
+            val alert = builder.create()
+            alert.show()
+
+
 
         }
